update build tasks/scripts/releasetools to GMIN specification
The way that Fastboot and EFI System Partition deliverables are
stored in the EFI system partition has changed. There are now
various zip and text files in the TFP under RADIO/ which have
the components to assemble these images. We have some new scripts
to create images from the target-files-package, as follows:

 bootloader_from_target_files: Create a VFAT EFI system partition
     image from the TFP

 fastboot_from_target_files: Create a fastboot boot image from
     the TFP

Both of these scripts live under releasetools/ as they are critical
for creating signed release artifacts from a TFP. All other scripts
now live in the root of this repo.

There is now a bootloader_from_zip script which is used by the build
system to create the ESP image, without a TFP as intermediary.

A new intel_common.py library contains frequently used functions
for creating fastboot and ESP images.

The generate_factory_images script has been extensively modified.
It now provides three scripts for partial or complete provisioning
of the device. There is now a new script included specifically for
CI runs that does not disturb any of the installed non-AOSP components.
A factory zip with testing keys is now populated with 'make dist'.

Makefiles and supportings scripts for the old way of updating the
EFI system partition have been removed

Depends-Change-Id: Id2e1eae6e707f97e4a1ed00c91d6d7073bc21312
Depends-Change-Id: I9f37836224b972430275932b50d96f869985dfc5
Depends-Change-Id: I83ae9b0129a085c30fdce4d0237397cf3e38ab7a
Change-Id: I7d4c99ba44dcc495704b58e69dbca5e09ef86b2b
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: 177502