Remove failure when lacking flashall target
We need to support multiple lunch targets, some of which will not
produce the flashall image target.

In those cases we will no longer error out.  Instead we will check to
see if the flashall image file will be produced, if it is we will update
it.  If not we will skip the step of updating the image file with the
flash.json file.

Change-Id: Iaf3a683677c514e00ae791d08584066913c1dfb0
Signed-off-by: John L. Villalovos <john.l.villalovos@intel.com>
Reviewed-on: 177796
Reviewed-by: Gumbel, Matthew K <matthew.k.gumbel@intel.com>