generate_factory_images: support fastboot args in individual scripts
The --fastboot-args option is removed. Its only use-case was to hardcode
some IP address information in the resulting scripts.

Now all the scripts will pass any additional command line arguments
passed in directly to the Fastboot command. We need this to support
DHCP ethernet scenarios where the device's IP address is not known
a priori.

Change-Id: I16c8f5823464ba46ed25d95d3838bae595febd7a
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/182782
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Phillips, Scott D <scott.d.phillips@intel.com>
Tested-by: cactus <cactus@intel.com>