sign_target_files_bootimages, sign_target_files_efis
sign_target_files_bootimages is used in production key
workflows to create prebuilt boot images in the destination
target-files-package that are signed with a production key.
Releasetools will favor the prebuilts if it finds them in the
archive.

sign_target_files_efis allows for re-signing EFI binaries that
go into the bootloader partition and the bootable Fastboot USB
image. It also can replace a built-in certificate in the shim,
or replace the current shim with a new one provided on the
command line.

- intel_common.GetFastbootImage() amended to communicate a
  password if supplied

- GetBootableImage() and BuildBootableImage() copied from AOSP
  releasetools and amended to communicate a password if supplied

- some extra slashes removed from pwfile that shouldn't be there

Depends-Change-Id: Ib01b2e826ca522db88f1d6af9bfa1f534eb72cc0
Change-Id: Ie7a92e91c6dca4804aae7584cc376e25b249630c
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: 185312
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Gumbel, Matthew K <matthew.k.gumbel@intel.com>