verify_from_target_files
This script accepts a path to a target-files-package as its only
parameter. It then communicates over the device and verifies the
hashes of all the boot images and files in the EFI System Partition
and /system.

We can use this script to assert that a device was provisioned or
updated OTA correctly.

Change-Id: I6d0a7e224fd66c8e27197d7f7c1bab8623d3be54
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/188757
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Bergeron, Michael <michael.bergeron@intel.com>