ota-test-prepare: allow for release testing, other minor fixes
The original version of this script generates 2 source builds and
generates necessary signed artifacts for running the OTA test cases.
However, when QA tests releases, they would prefer to test with
the current N and N-1 releases.

The -t command accepts a target-files zipfile as input and goes
through all the signing and image generation procedure with it.
In this way, this script can be invoked twice (with release N
and release N-1) to generate the A and B needed for the OTA test
cases.

Other minor fixes:

- Online help improved
- getopt options corrected
- Cleanup changed files in EFI prebuilt dirs

Change-Id: I84895b0cc5215d677b928a5be36a35f5e55e7b7b
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/190599
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Bergeron, Michael <michael.bergeron@intel.com>
Reviewed-by: Pepper, Timothy C <timothy.c.pepper@intel.com>