Add version of signapk that calls the DPD signing tool
This creates a signapk jar file, which instead of signing by itself,
calls the HW signing infrastructure to do the signature.  The private
key file is instead a text file with the certificate name that is
used to sign.  The SignFile binary itself is not included in this
commit.  Signapk now uses an environment variable SIGNFILE_PATH to
find the singing tool binary.

Change-Id: Ib98d8e2c5fca61ba3446b9047d4ca0207059b4e6
Signed-off-by: Matt Wood <matthew.d.wood@intel.com>
Signed-off-by: Michael Leibowitz <michael.leibowitz@intel.com>
Reviewed-on: https://android.intel.com/217100
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>
Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>