ota-test-prepare: detect proper concurrency level
The formula used is <num real cpus> * 2 + 1. This seems a reasonable
heuristic on most systems we have tested on. We don't count HT CPUs
as they aren't equivalent to a real core (typically only 20% increase
instead of 100% increase) and they are sufficiently leveraged by the * 2 
multiplier.

Supporting data, using a clean build of the default build target without
ccache on a Dual-CPU Xeon with 20 physical cores (40 with HT):

-j80: 18m10s
-j40: 18m12s
-j20: 21m31s

Change-Id: I42d35e80792be36a977371524c51dbeea40c54da
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: 284348
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>