[KF2] build system extensions for building EFI binaries
Previously EFI binaries had to be built by hand and checked into the repo
as prebuilts. The following is an attempt to integrate building EFI binaries
with the Android Build System (ABS).

Any module which should be built as an EFI binary should include efi.mk at
the top of their Android.mk to ensure the necessary defintions are brought
into the build system.

 LOCAL_EFI_KEY_PAIR: Signing key pair for EFI executables
 LOCAL_OBJCOPY_FLAGS: Extra flags to pass to objcopy when transforming the
   built .so file into a .efi executable.

It is assumed that all EFI executables will be using gnu-efi, we use gnu-efi's
linker scripts and C runtime stub.

We define BUILD_EFI_EXECUTABLE and BUILD_EFI_STATIC_BINARY to produce signed
.efi executables and static libraries respectively. Both of these hook into
the main ABS binary.mk, but set up the special compilation flags and steps needed.

Depends-Change-Id: Ib4fd00a1d0effe8e347382a922da101ce26bf696
Change-Id: I62dde2e9f56d457d6a5dfa0489fc00a7251dc75b
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/292639
Reviewed-by: Uyyala, Sridhar <sridhar.uyyala@intel.com>