Add support of PFT parameters in flashfiles.json
Adding new syntax for commands that require several files in argument.
Files are named $1, $2, ..., $N where N is the position of the file
in the "target" list.

$N will be replaced by PhoneFlashTool variables, created from the names
in the "target" list, in the same order :

"tool --option1 $1 --option2 $2"
target = [file1, file2]

will be transformed into :

"tool --option1 ${file1} --option2 ${file2}"

Change-Id: I4b9aef6d639b2c538f7d7b607012d70d815c774a
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-6517
Signed-off-by: Francois-Nicolas Muller <francois-nicolas.muller@intel.com>
Reviewed-on: https://android.intel.com/338501
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>
Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>