(REVERTME) IRDA-2327: Fix Boot Failed after upgrade from K to L
The boot manager entry Anroid-IA points to shim.efi which does not exist
in L so the firmware displays this error. This puts an extra copy of
Kernelflinger inside the bootloader partition image called shim.efi

This patch is a workaround to updating the EFI boot entries on upgrades.
That is tracked at GMINL-4267.

Issue: IRDA-2327

Change-Id: I5061f3a778ee5702c32535831ede6cfb97479316
Signed-off-by: Brad Geltz <brad.geltz@intel.com>
Reviewed-on: 342130
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>
Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>