Add support to track lunch target when fixing-up OTAs
The build fingerprint tracks both fish name and lunch. This patch reads
the lunch name from the device_mapping file and applies it to the
necessary portion of the build fingerprint.

This patch also adds support for checking the version in the device
mapping file which has been added and 'incremented' to 1.

Change-Id: Ic68958711ccdd9aec162c991c89a742e4352577a
Signed-off-by: Michael Bergeron <michael.bergeron@intel.com>
Reviewed-on: https://android.intel.com/346217
Reviewed-by: Betancourt, Eric S <eric.s.betancourt@intel.com>
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>
Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>