Fingerprint update logic fix
This change allows for 'cohol' lunch target devices with MR0 based
builds to correctly update when the pre-fish parameter is removed.

Moving forward with MR1 and the future, you'll need to either specify
the pre-fish name or the lunch target for MR0 to MR1 incrementals
to be created correctly.

Tested on 'coho' and 'cohol' lunch targets.

Issue: GMINL-8260

Change-Id: Icc2cc347cebc33cc8c1dee37598a6559744a02b9
Signed-off-by: Brad Geltz <brad.geltz@intel.com>
Reviewed-on: 358197
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>
Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>