bootloader policy: scripts and keys
The OAK and BPM are time based authenticated variables.  This patch
introduces two new keys pairs (the OEM MUST set its own values):

- odm is the key pair used to set the time-based authenticated
  variables (OAK and BPM).

- OAK is the key pair that is used during the Bootloader Policy
  challenge (over fastboot commands) to change the Bootloader Policy.

The OAK EFI variable contains the SHA256 hash of the OAK root
certificate (and not the OAK certificate itself in order to safe space
in the EFI variable storage space).

The generate_blpolicy_oemvars script generates flash-able
OEMVARS (with time-based authenticated option) file for the OAK and
BPM variables.

testkeys/blpolicy-delete.txt allows to delete the BPM and OAK EFI
variables.  It has been generated with a timestamp for 2050/1/1.

Tracked-On: OAM-1395
Change-Id: I81392fd63ae7576bfd91f177c9bad24a5e17aeba
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: 396539
Reviewed-on: 405333