blobstore: rewrite and add bootvars support
The existing blobstore implementation had some issues which were simpler
to resolve with a redesign:

- Adding new blob store types broke compatibility as it changed the size of
  data structures
- Hashes for efficient lookup weren't encoded in the data and had to be
  computed at runtime
- The codebase was very large considering what it really needed to do.

The docstring for blobstore.py describes the new data structure format in
detail.

In addition we add support for a new blobstore data type "bootvars" which
are added to the kernel command line on a per-board basis.

Issue: GMINL-7582, GMINL-5555
Change-Id: Ibfa5990e5b0669702bbb67c43d6174f31fc4286a
Depends-Change-Id: Ia7f6e37de74fc80b732c2f588c8e6ea77dc5f3a4
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/376586
Reviewed-by: Chouleur, Sylvain <sylvain.chouleur@intel.com>