generate fast flashfiles to ease testing.
with this patch, "make droid" will generate images in
PRODUCT_OUT/fast_flashfiles which can be flashed by phoneflashtool.

This is much faster than make flashfiles, which rely on OTA and
target files package, which are very slow and monothreaded.

The drawback is that OTA will not work after flashing this images

Tracked-On: OAM-1777
Change-Id: I4f1da90290b2fe2b0597b3d9cae9bbf6d9f3ad70
Signed-off-by: Jocelyn Falempe <jocelyn.falempe@intel.com>
Reviewed-on: 407609