flash-update: call uncrypt service before pre-recovery
Service pre-recovery now only reboots the platform into recovery OS.
If the update package is stored into /data and if this partiton is
encrypted, the service uncrypt must be called to produce the blockmap
file prior to calling pre-recovery service.
Uncrypt uses the configuration file /cache/recovery/uncrypt_file to
get the path of the zip package.

Change-Id: I45c89a8893dd32aae7471e3b8048d4530d031ef4
Signed-off-by: Leo Sartre <leox.sartre@intel.com>
Tracked-On: OAM-394
Reviewed-on: 399169