Make flashfiles_from_target_files script compliant with FLS
The environment variable ANDROID_PRODUCT_OUT is defined when building
the package. If the script flashfiles_from_target_files is called with
a package that was built in a different environment than the one the
script is executed in, this variable will not be compliant.
The location of the file build.prop containing this variable should be
evaluated each time the script is called in function of the argument
and should not be configurable build time through flashfiles.

On the contrary, the path to the tool FlsTool and other scripts should
not be based on some package properties but should be retrieved
instead directly from the target files package (in provdata).

Finally, needs to create the directory FASTBOOT_IMG_DIR if it does not
exist.

Change-Id: Ic952a136208282a3b0052a97bac9322d80de61cc
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-7289
Reviewed-on: https://android.intel.com:443/407182
Reviewed-on: https://android.intel.com:443/432309