sofia build: Phone Flash Tool fastboot flash support
fastboot configuration is built to extract fastboot flash files from FLS
files, and flash them using fastboot.

For each FLS file present in the flash file json description and that
has the key fasboot set to yes, two PFT commands are being added:
 - one to extract the fastboot binary from the FLS file to a temporary
   directory
 - one to flash the extracted binary with fasboot
Default command parameters are defined through a global.

On the opposite side, the timeout value for configurations other than
fastboot are now retrieved from configurations parameters instead of
commands and the function to add a configuration adds automatically
all the parameters.

Finally, this patch fixes the PFT "restrict" key that should be given
a table of config name strings rather than a simple string (
nevertheless both are supported, but the table is the official format
for "restrict").

Change-Id: Idf0d8b7b02cbf7d3b0d8c3a8989ea7033607c568
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-5957
Signed-off-by: Gabriel Touzeau <gabrielx.touzeau@intel.com>
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Reviewed-on: https://android.intel.com:443/424610