use fstab in the source when generating incremental OTA
This was needed after a change in edify_generator.py was made by google
which broke incremental update. After that change the fstab is used from
source tfp.
This change is also similar to change in tools/releasetools/ota_from_target_files.py
which does the same thing. this fixes the incremental OTA.

Tracked-On: https://jira01.devtools.intel.com/browse/OAM-1374
Change-Id: I764ec03b5d493460cd225a8479a0aa46e93df62c
Signed-off-by: Manish Regmi <manish.regmi@intel.com>
Reviewed-on: https://android.intel.com:443/406819