flash-update: add new option to apply update via sideload
Adb now offer the command: "reboot sideload-auto-reboot" that reboot
the device in sideload, waiting for the host to start transfering the
update package via sideload and when update is finished, reboot the
device.
Introduce the "-r" option to offer the possibility to use this method
to update the device using sideload. Default behavior is to proceed
like the fota app will do: push the package, call uncrypt service to
produce the blockmap, writting command to be excecuted to apply the
package then reboot to recovery.

Refactoring:
- number of options/parameters are growing, so use getopts to manage
them:
   $PACKAGE should now be preceded by "-p"
   $dutID   should now be preceded by "-s"
- when board is not encrypted or ota package is small, we never see
the property init.svc.uncrypt set to "started" so detect the creation
of the status_file rather than monitoring the property.

Example of use, updating via sideload:
./flash-update -s INV151700079 -p ota-update.zip -r

Change-Id: I474fe002384d9736e94fdc186e05006c81afb8fe
Signed-off-by: Leo Sartre <leox.sartre@intel.com>
Tracked-On: OAM-1665
Reviewed-on: 407048