Make ini flashfiles more generic
Current flashfiles.ini syntax is restricted to a defined set of fields
for each commands.
This patch allows to add generic fields into a commands section and each
field can contain a file variable wich will be interpreted.
Some fields like 'variant' are exceptions and are not converted to a
json one.

Change-Id: I8133bb8d1d012f1eb3b14550a71cd9d3445dbe27
Tracked-On: OAM-1344
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: 408886