Fix variants flashfiles generation for Broxton
Broxton has now variants and its ifwi files are variant specific
However these files are not used in "args" option of flashfile.ini so
they are not picked by the script.

This patch extend the variant files detection to all dynamic files used
in flashfiles.ini, not only the ones from "args".
This patch also :
 - fix a redundancy which appears in variant configurations
   in case of multiple output files
-  fix telephony oemvars files for variants: The shortname of files must
   not be lower-cased or the clean_config_parameters() function will not
   find the match and will delete the file parameter from
   configurations.

Change-Id: I3d9ba6ccaa89fab2c55a4f94e6755ee502eea780
Tracked-On: OAM-4301
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: 418913