flashfiles: generate file list by variant
If a variant does not have an ifwi file, it should not be a problem.
This patch generates the file list to be picked-up by looking at each
variant ini file if the file is needed.
For the additional files, we need to check that the specific variant IP
have not removed an additional file from the generic list.

Change-Id: I9d0f45d6898c028dc05c7c787485a0623a198eee
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-8083
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: https://android.intel.com:443/436822