blobstore: take fishname from TARGET_PRODUCT_FISHNAME
fishname from device_mapping does not necessarily correspond to the name
appended to the device name.
The reason is that fishname from dmap MUST be PRODUCT_DEVICE, because it
is used like this in ota_deployment_fixup script to replace the value of
ro.product.device.
So instead of this dmap entry, we use the fishname provided by build
variable TARGET_PRODUCT_FISHNAME which is the value that will
effectively be appended to the device name at runtime by autodetect.cpp

Change-Id: I73c7c7ed52b8629eb866b67d8a6af4fffc3eecc5
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-10138
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: https://android.intel.com:443/443912