[REVERTME] Revert "flash-update: Fix dutID lookup"
This reverts commit 33663196dbaf656ff6203d2c491940e5fcc6a27a.

"adb wait-for-device" will block until a device shows up and enumerate
as "device". When using "-r" switch, the device will enumerate as
"sideload" hence "adb wait-for-device" will never return.
Revert this patch that introduce the use of "adb wait-for-device" for
the old version that was polling using "adb device" command.
This patch should be reverted when a version of adb with the patch:
https://android-review.googlesource.com/#/c/184290/ is packaged into
phoneflashtool.
Revert jira: https://jira01.devtools.intel.com/browse/OAM-14698

Change-Id: I61178b46705b9d189c729d7baf770b2b07735d4d
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-14697
Signed-off-by: Leo Sartre <leox.sartre@intel.com>
Reviewed-on: https://android.intel.com:443/454379