ota_deployment_fixup: speedup OTA update contents
Instead of rewriting each OTA zip component one by one into the new zip
file we can copy the whole file and then modify the components we have
to.
This avoid the unzip/rezip process of each non touched components which
is the major part of the OTA zip file.

This reduces the process time for a full OTA package from 2m20 to 1m20.

Change-Id: I7744e7a864e0af029b1dbc1330aceceba325a923
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-15492
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: https://android.intel.com:443/456518