Add script to generate OTA files
This script can generate incremental, full or both OTA files for given
target files packages. Variants OTA files are also generated if the
platform supports it.
It also generates a metadata.json file which contains build informations
of the generated files, to be used by buildbot to publish the OTA files.

Tracked-On: OAM-16225
Change-Id: Ic49c92ef14dee8e44d649dce8be8a8a2c2c3e67b
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: 457091