Add dynamic check and log when compressing release image
It takes a quite a long time to compress release image if image
size is quite big, once time elapse of gzip process exceeds the
threshold, it will be treated as no reponse or zombie.

Add logic to check the completion of compressing release image
and print log if compressing is still in progress.

Signed-off-by: Zhong,Fangjian <fangjian.zhong@intel.com>