Fix the issue that bluetooth cannot work in release build
bluetooth package is resigned with a intel test key in the release
build. In this scenario, the test key used to sign the bluetooth
apk should be matched to the bluetooth key defined in the seinfo,
or zygote cannot set the right selinux domain for the bluetooth
process.
Create new keys for bluetooth and nfc, and define
PRODUCT_MAINLINE_SEPOLICY_DEV_CERTIFICATES with the intel test
keys path, so the build system can determine the keys it wants.

Tracked-On: OAM-132157
Signed-off-by: Zhenlong Z Ji <zhenlong.z.ji@intel.com>